x/lone/client/cli/query_customtx.go
package cli

import (
	"context"

	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/darxan/lone/x/lone/types"
	"github.com/spf13/cobra"
)

func CmdListCustomtx() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "list-customtx",
		Short: "list all customtx",
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			pageReq, err := client.ReadPageRequest(cmd.Flags())
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryAllCustomtxRequest{
				Pagination: pageReq,
			}

			res, err := queryClient.CustomtxAll(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}

func CmdShowCustomtx() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-customtx [index]",
		Short: "shows a customtx",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) (err error) {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			argIndex := args[0]

			params := &types.QueryGetCustomtxRequest{
				Index: argIndex,
			}

			res, err := queryClient.Customtx(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
